requirements.txt
matplotlib==3.7.2
numpy==1.25.1
opencv_python==4.9.0.80
h5py
Pillow==10.3.0
pytorch_msssim==1.0.0
scikit_learn==1.4.0
scipy==1.13.1
tqdm==4.65.0
einops==0.8.0
tensorflow==2.14.0
scikit-image==0.25.2